pdx206: rootdir: Fix correct type-f nfc config not being set for japense carrier models
In the previous implementation the model was being reported as "", Which was causing it to be set as unknown. Now the new implementation checks if there is mention of the specific japense model in the LTALABEL path and if it is, the model is set accordingly. Otherwise the following error code would happen: libnfc_nci: unknown opcode:0x21'

This implementation might be janky but works and was written with my limited knowledge of sh files.

Test: Take a logcat and "libnfc_nci: unknown opcode:0x21'" error does not occur, Nfc works as expected
Tested by: Nezhdan (on xda)